Revert many things after rebase

+++ src/Generating/PieceModifier.cpp
@@ -17,8 +17,6 @@ static const int SEED_OFFSET = 135 * 13;
 
 
 
-
-
 ////////////////////////////////////////////////////////////////////////////////
 /** A modifier which is pseudo-randomly replacing blocks to other types and metas. */
 class cPieceModifierRandomizeBlocks:
@@ -240,7 +238,7 @@ public:
 		{
 			if (m_BlocksToReplace.count(BlockTypes[i].Type()))
 			{
-				float BlockRnd = PieceNoise.IntNoise2DInRange(a_PieceRot, static_cast<int>(i), 0, m_AllWeights);
+				float BlockRnd = PieceNoise.IntNoise2DInRange(a_PieceRot, static_cast<int>(i), 0.0F, static_cast<float>(m_AllWeights));
 
 				int weightDelta = 0;
 				for (auto & blockToRnd : m_BlocksToRandomize)
@@ -253,7 +251,7 @@ public:
 						// Per block meta params
 						if (blockToRnd.m_MinMeta < blockToRnd.m_MaxMeta)
 						{
-							int BlockMetaRnd = std::clamp(static_cast<int>(PieceNoise.IntNoise2DInRange(a_PieceRot*2, static_cast<int>(i), blockToRnd.m_MinNoiseMeta, blockToRnd.m_MaxNoiseMeta)), blockToRnd.m_MinMeta, blockToRnd.m_MaxMeta);
+							int BlockMetaRnd = std::clamp(static_cast<int>(PieceNoise.IntNoise2DInRange(a_PieceRot*2, static_cast<int>(i), static_cast<float>(blockToRnd.m_MinNoiseMeta), static_cast<float>(blockToRnd.m_MaxNoiseMeta))), blockToRnd.m_MinMeta, blockToRnd.m_MaxMeta);
 							BlockMetas[i] = static_cast<unsigned char>(BlockMetaRnd);
 						}
 						else if ((blockToRnd.m_MaxMeta > -1) && (blockToRnd.m_MaxMeta == blockToRnd.m_MinMeta))
@@ -268,7 +266,7 @@ public:
 				// All blocks meta params
 				if (m_MaxMeta > m_MinMeta)
 				{
-					int BlockMetaRnd = std::clamp(static_cast<int>(PieceNoise.IntNoise2DInRange(a_PieceRot*2, static_cast<int>(i), m_MinNoiseMeta, m_MaxNoiseMeta)), m_MinMeta, m_MaxMeta);
+					int BlockMetaRnd = std::clamp(static_cast<int>(PieceNoise.IntNoise2DInRange(a_PieceRot * 2, static_cast<int>(i), static_cast<float>(m_MinNoiseMeta), static_cast<float>(m_MaxNoiseMeta))), m_MinMeta, m_MaxMeta);
 					BlockMetas[i] = static_cast<unsigned char>(BlockMetaRnd);
 				}
 				else if ((m_MaxMeta > -1) && (m_MaxMeta == m_MinMeta))
